add n classes to tree
