+++ .gitignore
@@ -146,3 +146,4 @@ skgrf/grf/
 
 .DS_Store
 .idea/
+*.svg
