+++ skgrf/ensemble/regressor.py
@@ -155,6 +155,7 @@ class GRFRegressor(BaseGRFForest, RegressorMixin):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_n_classes()
         return self
 
     def predict(self, X):
