+++ skgrf/tree/local_linear_regressor.py
@@ -227,6 +227,7 @@ class GRFTreeLocalLinearRegressor(BaseGRFTree, RegressorMixin):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_n_classes()
         return self
 
     def predict(self, X):
