+++ skgrf/tree/quantile_regressor.py
@@ -183,6 +183,7 @@ class GRFTreeQuantileRegressor(BaseGRFTree, RegressorMixin):
         )
         self._ensure_ptr()
         self._set_sample_weights(np.ones(len(X)))
+        self._set_n_classes()
         return self
 
     def predict(self, X):
